backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
aiokafka
mysql-connector-python
python-multipart
python-dotenv
websockets
